:recycle: [Refactor] 쿼리문 수정 (Member로 찾지 않고 Member의 PK값으로 조회하도록 변경)
