+++ src/main/java/kcs/funding/fundingboost/domain/service/GiftHubItemService.java
@@ -1,9 +1,9 @@
 package kcs.funding.fundingboost.domain.service;
 
 import static kcs.funding.fundingboost.domain.dto.response.GiftHubDto.createGiftHubDto;
+import static kcs.funding.fundingboost.domain.exception.ErrorCode.NOT_FOUND_GIFTHUB_ITEM;
 import static kcs.funding.fundingboost.domain.exception.ErrorCode.NOT_FOUND_ITEM;
 import static kcs.funding.fundingboost.domain.exception.ErrorCode.NOT_FOUND_MEMBER;
-import static kcs.funding.fundingboost.domain.exception.ErrorCode.NOT_FOUND_GIFTHUB_ITEM;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,7 +15,7 @@ import kcs.funding.fundingboost.domain.entity.GiftHubItem;
 import kcs.funding.fundingboost.domain.entity.Item;
 import kcs.funding.fundingboost.domain.entity.Member;
 import kcs.funding.fundingboost.domain.exception.CommonException;
-import kcs.funding.fundingboost.domain.repository.GiftHubItemRepository;
+import kcs.funding.fundingboost.domain.repository.GiftHubItem.GiftHubItemRepository;
 import kcs.funding.fundingboost.domain.repository.ItemRepository;
 import kcs.funding.fundingboost.domain.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
@@ -39,7 +39,7 @@ public class GiftHubItemService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new CommonException(NOT_FOUND_MEMBER));
 
-        List<GiftHubItem> giftHubItems = giftHubItemRepository.findGiftHubItemsByMember(member);
+        List<GiftHubItem> giftHubItems = giftHubItemRepository.findGiftHubItemsByMember(member.getMemberId());
 
         return giftHubItems.stream()
                 .map(giftHubItem -> createGiftHubDto(giftHubItem.getItem(), giftHubItem))
